\name{assoc.domains}
\alias{assoc.domains}

\title{Association measures between domains in multidimensional sequence analysis}

\description{
Computes various measures of association between dimensions of multidimensional sequence data.
}

\usage{
assoc.domains(dlist, names, djsa)
}

\arguments{
  \item{dlist}{A list of dissimilarity matrices or dist objects (see \code{\link{dist}}), with one element per dimension of the multidimensional sequence data}
  \item{names}{A character vector of the names of the dimensions of the multidimensional sequence data}
  \item{djsa}{A dissimilarity matrix or a dist object (see \code{\link{dist}}), corresponding to the distances between the multimdimensional sequences}
}

\references{
Piccarreta R. (2017). Joint Sequence Analysis: Association and Clustering, \emph{Sociological Methods and Research}, Vol. 46(2), 252-287.
}

\author{Nicolas Robette}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
  }
}

\keyword{Longitudinal characteristics}
\keyword{State sequences} 
