\name{symPLS}
\alias{symPLS}

\title{
Symmetric PLS
}
\description{
Computes symmetric PLS for two groups of continuous variables
}
\usage{
symPLS(a,b)
}
\arguments{
  \item{a}{data frame of the first group of continuous variables}
  \item{b}{data frame of the second group of continuous variables}
}

\author{Nicolas Robette, Xavier Bry}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
  }
}

\keyword{Longitudinal characteristics}
\keyword{State sequences} 
