\name{trajact}
\alias{trajact}
\docType{data}
\title{
Sample of employment histories
}
\description{
A data frame describing the employment status from age 14 to age 50. It's a sample of 500 interviewees from "Biographies et entourage" survey (INED, 2001).
}
\usage{data("trajact")}
\format{
  A data frame with 500 observations and 37 variables. Each variable is numeric and describes the employment status at a given age : 1 = education, 2 = full-time job, 3 = part-time job, 4 = small jobs, 5 = inactivity, 6 = military service.
}

\examples{
data(trajact)
str(trajact)
}
\keyword{datasets}
