\name{symPLS}
\alias{symPLS}

\title{
Symmetric (or canonical) PLS
}
\description{
Computes symmetric (or canonical) PLS for two groups of continuous variables
}
\usage{
symPLS(a,b)
}
\arguments{
  \item{a}{data frame of the first group of continuous variables}
  \item{b}{data frame of the second group of continuous variables}
}

\author{Nicolas Robette, Xavier Bry}

\references{
Bry X. (1996). Analyses Factorielles Multiples. Paris, Economica Poche.
de Jong S., Wise B.M. and Ricker N.L. (2001). Canonical Partial Least Squares and Continuum Power Regression. \emph{Journal of Chemometrics}, Vol. 15, 85–100.
}

\keyword{Longitudinal characteristics}
\keyword{State sequences} 
